+++ src/bullet/index.tsx
@@ -22,10 +22,20 @@ const StyledBullet = styled(Bullet)`
   }
 
   &.kirk-bullet--map-active,
+  &.kirk-bullet--map-inactive,
+  &.kirk-bullet--search {
+    border-width: 3px;
+  }
+
+  &.kirk-bullet--map-active,
   &.kirk-bullet--map-inactive {
     width: ${componentSizes.bulletSizeMap};
     height: ${componentSizes.bulletSizeMap};
-    border-width: 3px;
+  }
+
+  &.kirk-bullet--search {
+    width: ${componentSizes.bulletSizeSearch};
+    height: ${componentSizes.bulletSizeSearch};
   }
 
   &.kirk-bullet--map-active {
@@ -36,6 +46,11 @@ const StyledBullet = styled(Bullet)`
   &.kirk-bullet--map-inactive {
     border-color: ${color.polylineSecondary};
     box-shadow: 0 0 0 1px ${color.polylineStrokeSecondary};
+  }
+
+  &.kirk-bullet--search {
+    border-width: 4px;
+    border-color: ${color.secondaryText};
   }
 `
 export { BulletTypes, BulletProps } from './Bullet'