+++ src/searchForm/autoComplete/section/index.tsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components'
+import { autoCompleteBaseStyle } from 'searchForm/baseStyles'
+
+import autoComplete from './AutoCompleteSection'
+
+const StyledAutoComplete = styled(autoComplete)`
+  & {
+    box-sizing: border-box;
+
+    ${autoCompleteBaseStyle}
+  }
+`
+
+export default StyledAutoComplete